Fix: change all needed int and serial in coverstore directory
- Change all needeed int and serail to bigint and bigserail in this files:
	openlibrary/coverstore
		- schema.py
		- schema.sql

+++ openlibrary/coverstore/schema.py
@@ -8,14 +8,14 @@ def get_schema(engine='postgres'):
 
     s.add_table(
         'category',
-        s.column('id', 'serial', primary_key=True),
+        s.column('id', 'bigserial', primary_key=True),
         s.column('name', 'string'),
     )
 
     s.add_table(
         'cover',
-        s.column('id', 'serial', primary_key=True),
-        s.column('category_id', 'integer', references='category'),
+        s.column('id', 'bigserial', primary_key=True),
+        s.column('category_id', 'bigint', references='category'),
         s.column('olid', 'string'),
         s.column('filename', 'string'),
         s.column('filename_s', 'string'),
@@ -41,8 +41,8 @@ def get_schema(engine='postgres'):
 
     s.add_table(
         "log",
-        s.column("id", "serial", primary_key=True),
-        s.column("cover_id", "integer", references="cover"),
+        s.column("id", "bigserial", primary_key=True),
+        s.column("cover_id", "bigint", references="cover"),
         s.column("action", "text"),
         s.column("timestamp", "timestamp"),
     )
